Create NAB & DA.ino
Title: Merging National Anthem Tones of Bangladesh with Delivery Animation

Objective: The purpose of this task is to merge the musical tones of the national anthem of Bangladesh, "Amar Shonar Bangla," with an accompanying visual delivery animation. This will create a synchronized multimedia file that can be used for ceremonial presentations, official functions, or other relevant uses.